Refactor analyze.py: Extract fitting utilities and CLI parser
Extract reusable components from the monolithic analyze.py (5108 lines)
into focused, well-documented modules:

- fitting_utils.py (140 lines): Utilities for working with fit results,
  parameter extraction, efficiency resolution, and uncertainty calculations
- cli_parser.py (334 lines): Complete command-line argument parsing logic

Benefits:
- Reduced analyze.py from 5108 to 4658 lines (-450 lines, -9%)
- Improved code organization and discoverability
- Made fitting utilities reusable across modules
- Separated CLI concerns from analysis logic
- Better testability for isolated components

All functions maintain identical behavior; this is a pure refactoring with
no functional changes.
